refactor(tui): carve HF-pull subsystem out of events.rs into hf_pull.rs
Move the self-contained HuggingFace-pull / download behavior (~24 fns +
the StripProgress DownloadProgress impl + its tests) out of the 5512-line
events.rs god-file into src/tui/hf_pull.rs (events.rs 5512 -> 4651). The
state half already lived in tui::hf_dialog / tui::download_strip; this is
their behavior half. Verbatim move, visibility-only changes, no behavior
change. The optional input/actions/tasks three-way split stays deferred.
